add status

+++ src/getStatus.h
@@ -5,7 +5,7 @@
 
 
 #include <iostream>
-
+#include"act_d435.h"
 #ifdef __linux__
 #include <unistd.h>
 #include "serial.h"
@@ -13,7 +13,7 @@
 void GetStatus(serial::Serial *my_serial,unsigned int *runStatus,int *playGround);
 
 void UpdateStatus(serial::Serial *my_serial,unsigned int *runStatus,int *playGround);
-void SendDatas(serial::Serial *my_serial, int status, double frountData, double lateralDatas);
+void SendDatas(serial::Serial *my_serial, int status, float frountData, float lateralDatas);
 void CopyData(uint8_t* origen, uint8_t* afterTreat, int size)
 union dataUnion
 {
